chore: migrate parsing logic of filter to filter service
