+++ src/JsonApiDotNetCore/Services/QueryParser.cs
@@ -57,7 +57,7 @@ namespace JsonApiDotNetCore.Services
                 if (pair.Key.StartsWith(QueryConstants.FILTER, StringComparison.Ordinal))
                 {
                     if (disabledQueries.HasFlag(QueryParams.Filters) == false)
-                        //querySet.Filters.AddRange(ParseFilterQuery(pair.Key, pair.Value));
+                        _filterService.Parse(pair.Key, pair.Value);
                     continue;
                 }
 
@@ -103,49 +103,6 @@ namespace JsonApiDotNetCore.Services
                 .Select(t => (IQueryParameterService)_sp.GetService(t));
         }
 
-        protected virtual List<FilterQuery> ParseFilterQuery(string key, string value)
-        {
-            // expected input = filter[id]=1
-            // expected input = filter[id]=eq:1
-            var queries = new List<FilterQuery>();
-            var propertyName = key.Split(QueryConstants.OPEN_BRACKET, QueryConstants.CLOSE_BRACKET)[1];
-
-            // InArray case
-            string op = GetFilterOperation(value);
-            if (string.Equals(op, FilterOperations.@in.ToString(), StringComparison.OrdinalIgnoreCase)
-                || string.Equals(op, FilterOperations.nin.ToString(), StringComparison.OrdinalIgnoreCase))
-            {
-                (var operation, var filterValue) = ParseFilterOperation(value);
-                queries.Add(new FilterQuery(propertyName, filterValue, op));
-            }
-            else
-            {
-                var values = value.Split(QueryConstants.COMMA);
-                foreach (var val in values)
-                {
-                    (var operation, var filterValue) = ParseFilterOperation(val);
-                    queries.Add(new FilterQuery(propertyName, filterValue, operation));
-                }
-            }
-
-            return queries;
-        }
-
-        protected virtual (string operation, string value) ParseFilterOperation(string value)
-        {
-            if (value.Length < 3)
-                return (string.Empty, value);
-
-            var operation = GetFilterOperation(value);
-            var values = value.Split(QueryConstants.COLON);
-
-            if (string.IsNullOrEmpty(operation))
-                return (string.Empty, value);
-
-            value = string.Join(QueryConstants.COLON_STR, values.Skip(1));
-
-            return (operation, value);
-        }
 
         protected virtual PageQuery ParsePageQuery(PageQuery pageQuery, string key, string value)
         {
@@ -216,19 +173,6 @@ namespace JsonApiDotNetCore.Services
             }
         }
 
-        private string GetFilterOperation(string value)
-        {
-            var values = value.Split(QueryConstants.COLON);
 
-            if (values.Length == 1)
-                return string.Empty;
-
-            var operation = values[0];
-            // remove prefix from value
-            if (Enum.TryParse(operation, out FilterOperations op) == false)
-                return string.Empty;
-
-            return operation;
-        }
     }
 }
